roll back to using legacy routes
