+++ src/renderer.tsx
@@ -37,13 +37,8 @@ import { Provider } from 'react-redux'
 
 import PastelDB from './features/pastelDB/database'
 import { fetchPastelPrice } from './features/pastelPrice'
-import Routes from './common/routes/Routes'
+import Root from './legacy/containers/Root'
 import store from './redux/store'
-import { ThemeProvider } from 'styled-components/macro'
-import { theme } from './common/theme'
-import { MemoryRouter } from 'react-router-dom'
-import { PastelModal } from './features/pastelModal'
-import UpdateToast from './features/updateToast'
 
 const oneHour = 1000 * 60 * 60
 /**
@@ -70,13 +65,7 @@ try {
 
 const application = (
   <Provider store={store}>
-    <ThemeProvider theme={theme}>
-      <MemoryRouter>
-        <Routes />
-        <PastelModal />
-        <UpdateToast />
-      </MemoryRouter>
-    </ThemeProvider>
+    <Root />
   </Provider>
 )
